Fix message counter and remove vulnerability (read desc)
Vulnerability consisted being able to modify attributes of the flask
backend from the outside. When there would be a number in the message,
not necessarily marking the sender's number, but e.g. someone who sent a
message with someone else's number, that other number gets added to the
messages counter and may get blocked. This vulenerability used the "in"
keyword to check for numbers in the message, now we use the find
function to check if the number is at the start of the message, which is
always is unless the Automate flow is modified. Another thing this
vulnerability would likely cause is different sender's messages being
mixed in the sender_chat_log, which would have the potential of leaking
some information from other chats.